Refactor: Enhance Code Structure Using Mixins for Configuring Ellipsis Actions and Handling Actions
