+++ resources/js/processes/components/ProcessesListing.vue
@@ -47,7 +47,7 @@
         <template slot="actions" slot-scope="props">
           <ellipsis-menu 
             @navigate="onNavigate"
-            :actions="actions"
+            :actions="processActions"
             :permission="permission"
             :data="props.rowData"
             :is-documenter-installed="isDocumenterInstalled"
@@ -59,12 +59,12 @@
       <create-pm-block-modal id="create-pm-block-modal" ref="create-pm-block-modal" :currentUserId="currentUserId" :assetName="pmBlockName" :assetId="processId" />
       <add-to-project-modal id="add-to-project-modal" ref="add-to-project-modal"  assetType="process" :assetId="processId" :assetName="assetName"/>
       <pagination
-              :single="$t('Process')"
-              :plural="$t('Processes')"
-              :per-page-select-enabled="true"
-              @changePerPage="changePerPage"
-              @vuetable-pagination:change-page="onPageChange"
-              ref="pagination"
+        :single="$t('Process')"
+        :plural="$t('Processes')"
+        :per-page-select-enabled="true"
+        @changePerPage="changePerPage"
+        @vuetable-pagination:change-page="onPageChange"
+        ref="pagination"
       ></pagination>
     </div>
   </div>
@@ -79,31 +79,18 @@
   import CreateTemplateModal from "../../components/templates/CreateTemplateModal.vue";
   import CreatePmBlockModal from "../../components/pm-blocks/CreatePmBlockModal.vue";
   import EllipsisMenu from "../../components/shared/EllipsisMenu.vue";
+  import ellipsisMenuMixin from "../../components/shared/ellipsisMenuActions";
   import AddToProjectModal from "../../components/shared/AddToProjectModal.vue";
+  import processNavigationMixin from "../../components/shared/processNavigation";
 
   const uniqIdsMixin = createUniqIdsMixin();
 
   export default {
     components: { TemplateExistsModal, CreateTemplateModal, EllipsisMenu, CreatePmBlockModal, AddToProjectModal},
-    mixins: [datatableMixin, dataLoadingMixin, uniqIdsMixin],
+    mixins: [datatableMixin, dataLoadingMixin, uniqIdsMixin, ellipsisMenuMixin, processNavigationMixin],
     props: ["filter", "id", "status", "permission", "isDocumenterInstalled", "pmql", "processName", "currentUserId"],
     data() {
       return {
-        actions: [
-        { value: "unpause-start-timer", content: "Unpause Start Timer Events", icon: "fas fa-play", conditional: "if(has_timer_start_events and pause_timer_start, true, false)" },
-        { value: "pause-start-timer", content: "Pause Start Timer Events", icon: "fas fa-pause", conditional: "if(has_timer_start_events and not(pause_timer_start), true, false)"},
-        { value: "edit-designer", content: "Edit Process", link: true, href:"/modeler/{{id}}", permission: "edit-processes", icon: "fas fa-edit", conditional: "if(status == 'ACTIVE' or status == 'INACTIVE', true, false)"},
-        { value: "create-template", content: "Save as Template", permission: "create-process-templates", icon: "fas fa-layer-group" },
-        { value: "create-pm-block", content: "Save as PM Block", permission: "create-pm-blocks", icon: "fas nav-icon fa-cube" },
-        { value: "add-to-project", content: "Add to Project", icon: "fas fa-folder-plus" },
-        { value: "edit-item", content: "Configure", link: true, href:"/processes/{{id}}/edit", permission: "edit-processes", icon: "fas fa-cog", conditional: "if(status == 'ACTIVE' or status == 'INACTIVE', true, false)"},
-        { value: "view-documentation", content: "View Documentation", link: true, href:"/modeler/{{id}}/print", permission: "view-processes", icon: "fas fa-sign", conditional: "isDocumenterInstalled"},
-        { value: "remove-item", content: "Archive", permission: "archive-processes", icon: "fas fa-archive", conditional: "if(status == 'ACTIVE' or status == 'INACTIVE', true, false)" },
-        { value: "divider" },
-        { value: "export-item", content: "Export", link: true, href:"/processes/{{id}}/export", permission: "export-processes", icon: "fas fa-file-export"},
-        { value: "restore-item", content: "Restore", permission: "archive-processes", icon: "fas fa-upload", conditional: "if(status == 'ARCHIVED', true, false)" },
-        { value: "download-bpmn", content: "Download BPMN", permission: "view-processes", icon: "fas fa-file-download"},
-      ],
         orderBy: "name",
         processId: null,
         processTemplateName: '',
@@ -176,106 +163,8 @@
       showAddToProjectModal(name, id) {        
         this.processId = id;
         this.assetName = name;
+        this.assetType = "process";
         this.$refs["add-to-project-modal"].show();
-      },
-      onNavigate(action, data) {
-        let putData = {
-          name: data.name,
-          description: data.description,
-        };
-        switch (action.value) {
-          case "unpause-start-timer":
-            putData.pause_timer_start = false;
-            ProcessMaker.apiClient
-                .put("processes/" + data.id, putData)
-                .then(response => {
-                  ProcessMaker.alert(
-                      this.$t("The process was unpaused."),
-                      "success"
-                  );
-                  this.$emit("reload");
-                });
-            break;
-          case "pause-start-timer":
-            putData.pause_timer_start = true;
-            ProcessMaker.apiClient
-                .put("processes/" + data.id, putData)
-                .then(response => {
-                  ProcessMaker.alert(
-                      this.$t("The process was paused."),
-                      "success"
-                  );
-                  this.$emit("reload");
-                });
-            break;
-          case "create-template":
-            this.showCreateTemplateModal(data.name, data.id);
-            break;
-          case "create-pm-block":
-            this.showPmBlockModal(data.name, data.id);
-            break;
-          case "restore-item":
-            ProcessMaker.apiClient
-                .put("processes/" + data.id + "/restore")
-                .then(response => {
-                  ProcessMaker.alert(
-                      this.$t("The process was restored."),
-                      "success"
-                  );
-                  this.$emit("reload");
-                });
-            break;
-          case "remove-item":
-            ProcessMaker.confirmModal(
-                this.$t("Caution!"),
-                this.$t("Are you sure you want to archive the process") +
-                data.name +
-                "?",
-                "",
-                () => {
-                  ProcessMaker.apiClient
-                      .delete("processes/" + data.id)
-                      .then(response => {
-                        ProcessMaker.alert(
-                            this.$t("The process was archived."),
-                            "success"
-                        );
-                        this.$refs.pagination.loadPage(1);
-                      });
-                }
-            );
-            break;
-
-          case "download-bpmn":
-            ProcessMaker.confirmModal(
-                this.$t("Caution!"),
-                this.$t("Are you sure you want to download the BPMN definition of the process?"),
-                "",
-                () => {
-                  ProcessMaker.apiClient
-                      .get("processes/" + data.id + "/bpmn")
-                      .then(response => {
-
-                        const link = document.createElement("a");
-                        const file = new Blob([response.data], { type: 'text/plain' });
-
-                        link.href = URL.createObjectURL(file);
-                        link.download = "bpmnProcess.xml";
-                        link.click();
-                        URL.revokeObjectURL(link.href);
-
-                        ProcessMaker.alert(
-                            this.$t("The process BPMN has been downloaded."),
-                            "success"
-                        );
-                      });
-                }
-            );
-            break;
-            case 'add-to-project':
-              this.showAddToProjectModal(data.name, data.id);
-            break;
-        }
       },
       formatStatus(status) {
         status = status.toLowerCase();
