+++ resources/js/processes/scripts/components/ScriptListing.vue
@@ -31,7 +31,7 @@
 
         <template slot="actions" slot-scope="props">
           <ellipsis-menu
-            :actions="actions"
+            :actions="scriptActions"
             :permission="permission"
             :data="props.rowData"
             :divider="true"
@@ -89,53 +89,19 @@
 import datatableMixin from "../../../components/common/mixins/datatable";
 import dataLoadingMixin from "../../../components/common/mixins/apiDataLoading";
 import EllipsisMenu from "../../../components/shared/EllipsisMenu.vue";
+import ellipsisMenuMixin from "../../../components/shared/ellipsisMenuActions";
 import AddToProjectModal from "../../../components/shared/AddToProjectModal.vue";
 import { createUniqIdsMixin } from "vue-uniq-ids";
 const uniqIdsMixin = createUniqIdsMixin();
 
 export default {
   components: { EllipsisMenu, AddToProjectModal },
-  mixins: [datatableMixin, dataLoadingMixin, uniqIdsMixin],
+  mixins: [datatableMixin, dataLoadingMixin, uniqIdsMixin, ellipsisMenuMixin],
   props: ["filter", "id", "permission", "scriptExecutors"],
   data() {
     return {
       assetId: null,
       assetName: "",
-      actions: [
-        {
-          value: "edit-script",
-          content: "Edit Script",
-          link: true,
-          href: "/designer/scripts/{{id}}/builder",
-          permission: "edit-scripts",
-          icon: "fas fa-pen-square",
-        },
-        {
-          value: "edit-item",
-          content: "Configure",
-          link: true,
-          href: "/designer/scripts/{{id}}/edit",
-          permission: "edit-scripts",
-          icon: "fas fa-cog",
-        },
-        { 
-          value: "add-to-project", 
-          content: "Add to Project",
-          icon: "fas fa-folder-plus" 
-        },
-        {
-          value: "duplicate-item",
-          content: "Copy",
-          permission: "create-scripts",
-          icon: "fas fa-copy",
-        },
-        {
-          value: "remove-item",
-          content: "Delete",
-          permission: "delete-scripts",
-          icon: "fas fa-trash-alt",
-        },
-      ],
       dupScript: {
         title: "",
         type: "",
